ScriptPlayer component added @TGA-008
