+++ src/components/Animation/index.js
@@ -1,39 +1,24 @@
 import * as THREE from 'three';
 
 
-export const Animation = (
+export const Animation = ({
   gltf,
-  { clipActions,
-    scripts
+  clipActions
   }) => {
   const animationMixer = new THREE.AnimationMixer(gltf.scene);
   const animationClips = gltf.animations;
   const clipActionsMap = new Map();
   let previousAction;
   let activeAction;
-  let scriptState = {};
-  let rotateFlag = 0;
   let clipStartTime = 0;
+  let animationMixerFinishedCallback;
 
   animationClips.filter(ac => clipActions.includes(ac.name)).forEach(ac => clipActionsMap.set(ac.name, animationMixer.clipAction(ac)));
-  animationMixer.addEventListener('finished', clipActionFinished);
 
-  // START NEW SCRIPT
-  function startNewScript (scriptName) {
-    scriptState = {
-      scriptName,
-      clipNames: scripts[scriptName],
-      scriptLength: scripts[scriptName].length,
-      clipIdx: 0
-    }
-    playNextClipActionInScript();
-  }
-
-  function playNextClipActionInScript () {
-    if (scriptState.scriptName === undefined) { console.log('scriptState.scriptName === undefined'); return; }
-    const { clipIdx, clipNames } = scriptState;
-    const { clipName } = clipNames[clipIdx];
-    playClipAction(clipName); 
+  const setAnimationMixerFinishedCallback = (newAnimationMixerFinishedCallback) => {
+    animationMixer.removeEventListener('finished', animationMixerFinishedCallback);
+    animationMixer.addEventListener('finished', newAnimationMixerFinishedCallback);
+    animationMixerFinishedCallback = newAnimationMixerFinishedCallback;
   }
 
   const playClipAction = (clipName) => {
@@ -50,25 +35,13 @@ export const Animation = (
       clipStartTime = Date.now();
   }
 
-  function clipActionFinished (ev) {
-    const { clipIdx, clipNames, scriptLength } = scriptState;
-    const { clipName, loop, rotate } = clipNames[clipIdx];
-    rotateFlag = rotate;
-    scriptState.clipIdx = (clipIdx + 1) % scriptLength;
-    playNextClipActionInScript();
-  }
-
-
-  const resetRotateFlag = () => rotateFlag = 0;
 
   const update = (deltaSeconds) => animationMixer.update(deltaSeconds);
   
   return {
     get clipStartTime() { return clipStartTime },
-    get rotateFlag() { return rotateFlag },
-    resetRotateFlag,
-    startNewScript,
     playClipAction,
+    setAnimationMixerFinishedCallback,
     update
   }
