+++ src/components/Character/index.js
@@ -4,6 +4,7 @@ import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader";
 import { Animation } from '../Animation';
 import { Rotation } from '../Rotation';
 import { Translation } from '../Tanslation';
+import { ScriptPlayer } from '../ScriptPlayer';
 
 
 export function Character({ 
@@ -20,8 +21,10 @@ export function Character({
   let animationMixer;
   // ROTATION HANDLER
   let rotation = Rotation({ mesh, defaultRotation: 0 });
-  // TRANLATIONS HANDLER
+  // TRANSLATIONS HANDLER
   let translation = Translation({ mesh, defaultPosition: [0, 0, 0] });
+  // SCRIPT PLAYER
+  let scriptPlayer;
 
   gltfLoader.load(assetPath, (gltf) => {
     gltf.scene.scale.set(meshScaler, meshScaler, meshScaler);
@@ -29,8 +32,9 @@ export function Character({
     const model = gltf.scene;
     model.position.set(0, 0, 0);
     mesh.add(model);
-    animationMixer = Animation(gltf, ANIMATION_CONFIGS);
-    animationMixer.startNewScript('idle');
+    animationMixer = Animation({ gltf, clipActions: ANIMATION_CONFIGS.clipActions });
+    scriptPlayer = ScriptPlayer({ animationMixer, scripts: ANIMATION_CONFIGS.scripts });
+    scriptPlayer.startNewScript('idle');
     if (onLoadCallback !== undefined) onLoadCallback();
   });
 
@@ -41,11 +45,11 @@ export function Character({
     // UPDATE ANIMATION
     animationMixer?.update(deltaSeconds);
     // CHECK FOR ROTATION
-    if (animationMixer?.rotateFlag !== 0) {
+    if (scriptPlayer?.rotateFlag !== 0) {
       const clipTime = Date.now() - animationMixer?.clipStartTime;
       if (clipTime > 1) { 
-        rotation.y = animationMixer?.rotateFlag;
-        animationMixer.resetRotateFlag()
+        rotation.y = scriptPlayer?.rotateFlag;
+        scriptPlayer.resetRotateFlag()
       }
     }
     // CHECK FOR TRANSLATION
@@ -59,7 +63,6 @@ export function Character({
     get visible() { return mesh.visible },
     get matrix() { return mesh.matrix },
     setPosition: (newPos) => translation.setPosition(newPos),
-    update,
-    //setMatrixFromArray
+    update
   }
 }
